feat: add 404 handling
