+++ product-service/constants.js
@@ -1,0 +1,60 @@
+const products = [
+    {
+      "count": 4,
+      "description": "Short Product Description111111",
+      "id": "7567ec4b-b10c-48c5-9345-fc73c48a80aa",
+      "price": 2.4,
+      "title": "ProductOneFromSLS111111"
+    },
+    {
+      "count": 6,
+      "description": "Short Product Description3",
+      "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a0",
+      "price": 10,
+      "title": "ProductNewFromSLS"
+    },
+    {
+      "count": 7,
+      "description": "Short Product Description2",
+      "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a2",
+      "price": 23,
+      "title": "ProductTopFromSLS"
+    },
+    {
+      "count": 12,
+      "description": "Short Product Description7",
+      "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a1",
+      "price": 15,
+      "title": "ProductTitleFromSLS"
+    },
+    {
+      "count": 7,
+      "description": "Short Product Description2",
+      "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a3",
+      "price": 23,
+      "title": "ProductFromSLS"
+    },
+    {
+      "count": 8,
+      "description": "Short Product Description4",
+      "id": "7567ec4b-b10c-48c5-9345-fc73348a80a1",
+      "price": 15,
+      "title": "ProductTestFromSLS"
+    },
+    {
+      "count": 2,
+      "description": "Short Product Descriptio1",
+      "id": "7567ec4b-b10c-48c5-9445-fc73c48a80a2",
+      "price": 23,
+      "title": "Product2FromSLS"
+    },
+    {
+      "count": 3,
+      "description": "Short Product Description7",
+      "id": "7567ec4b-b10c-45c5-9345-fc73c48a80a1",
+      "price": 15,
+      "title": "ProductNameFromSLS"
+    }
+  ]
+
+  exports.products = products;
