+++ product-service/handler.js
@@ -1,70 +1,13 @@
 'use strict';
 
-// import products from './products.json' assert {type: 'json'};
-const products = [
-  {
-    "count": 4,
-    "description": "Short Product Description1",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80aa",
-    "price": 2.4,
-    "title": "ProductOneFromSLS"
-  },
-  {
-    "count": 6,
-    "description": "Short Product Description3",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a0",
-    "price": 10,
-    "title": "ProductNewFromSLS"
-  },
-  {
-    "count": 7,
-    "description": "Short Product Description2",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a2",
-    "price": 23,
-    "title": "ProductTopFromSLS"
-  },
-  {
-    "count": 12,
-    "description": "Short Product Description7",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a1",
-    "price": 15,
-    "title": "ProductTitleFromSLS"
-  },
-  {
-    "count": 7,
-    "description": "Short Product Description2",
-    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a3",
-    "price": 23,
-    "title": "ProductFromSLS"
-  },
-  {
-    "count": 8,
-    "description": "Short Product Description4",
-    "id": "7567ec4b-b10c-48c5-9345-fc73348a80a1",
-    "price": 15,
-    "title": "ProductTestFromSLS"
-  },
-  {
-    "count": 2,
-    "description": "Short Product Descriptio1",
-    "id": "7567ec4b-b10c-48c5-9445-fc73c48a80a2",
-    "price": 23,
-    "title": "Product2FromSLS"
-  },
-  {
-    "count": 3,
-    "description": "Short Product Description7",
-    "id": "7567ec4b-b10c-45c5-9345-fc73c48a80a1",
-    "price": 15,
-    "title": "ProductNameFromSLS"
-  }
-]
+const constants = require('./constants');
 
 module.exports.getProductsList = async (event) => {
+  
   return {
     statusCode: 200,
     body: JSON.stringify(
-      products,
+      constants.products,
       null,
       2
     ),
@@ -77,15 +20,32 @@ module.exports.getProductsList = async (event) => {
 };
 
 module.exports.getProductsById = async (event) => {
-  const product = products.filter((el) => el.id === event.pathParameters.id)
-  return {
-    statusCode: 200,
-    body: JSON.stringify(
-      product,
-      null,
-      2
-    ),
-  };
+  const product = constants.products.filter((el) => el.id === event.pathParameters.id)
+  try {
+    if (product.length === 0) {
+      throw new Error()
+    } else {
+      return {
+        statusCode: 200,
+        body: JSON.stringify(
+          product,
+          null,
+          2
+        ),
+      };
+    }
+  }
+  catch(err) {
+    return {
+      statusCode: 404,
+      body: JSON.stringify(
+        "Product not found",
+        null,
+        2
+      ),
+    };
+  }
+
 
   // Use this code if you don't use the http event with the LAMBDA-PROXY integration
   // return { message: 'Go Serverless v1.0! Your function executed successfully!', event };
