+++ product-service/package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "serverless-webpack": "^5.9.1"
+  }
+}
